:recycle: pages層から表示・計算ロジックを剥がす
